Mvvm

+++ core/domain/src/main/java/com/example/domain/models/Content.kt
@@ -11,7 +11,7 @@ sealed class Content {
         val url: String,
         val photographer: String,
         val photographerUrl: String,
-        val photographerId: Int,
+        val photographerId: Long,
         val avgColor: String,
         val src: Src,
         val liked: Boolean,
