+++ core/pexelsApi/src/main/java/com/example/pexels_api/models/PhotoDto.kt
@@ -39,7 +39,7 @@ data class PhotoDto(
     val photographerUrl: String?,
 
     @SerialName("photographer_id")
-    val photographerId: Int?,
+    val photographerId: Long?,
 
     @SerialName("avg_color")
     val avgColor: String?,
